Fixed controllers, updated UI

+++ ScientificReport/Controllers/ReportThesisController.cs
@@ -2,7 +2,6 @@ using System;
 using System.Linq;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 using ScientificReport.BLL.Interfaces;
 using ScientificReport.Controllers.Utils;
 using ScientificReport.DAL.Entities;
@@ -16,22 +15,31 @@ namespace ScientificReport.Controllers
 	{
 		private readonly IReportThesisService _reportThesisService;
 		private readonly IUserProfileService _userProfileService;
+		private readonly IDepartmentService _departmentService;
+		private readonly IConferenceService _conferenceService;
 
-		public ReportThesisController(IReportThesisService reportThesisService, IUserProfileService userProfileService)
+		public ReportThesisController(
+			IReportThesisService reportThesisService,
+			IUserProfileService userProfileService,
+			IDepartmentService departmentService,
+			IConferenceService conferenceService
+		)
 		{
 			_reportThesisService = reportThesisService;
 			_userProfileService = userProfileService;
+			_departmentService = departmentService;
+			_conferenceService = conferenceService;
 		}
 
 		// GET: ReportThesis
 		public IActionResult Index(ReportThesisIndexModel model)
 		{
-			model.ReportTheses = _reportThesisService.GetPage(model.CurrentPage, model.PageSize);
-			model.Count = _reportThesisService.GetCount();
+			model.ReportTheses = _reportThesisService.GetPageByRole(model.CurrentPage, model.PageSize, User);
+			model.Count = _reportThesisService.GetCountByRole(User);
 			return View(model);
 		}
 
-		// GET: ReportThesis/Details/5
+		// GET: ReportThesis/Details/{id}
 		public IActionResult Details(Guid? id)
 		{
 			if (id == null)
@@ -46,6 +54,11 @@ namespace ScientificReport.Controllers
 				return NotFound();
 			}
 
+			if (!UserHasPermission(reportThesis))
+			{
+				return Forbid();
+			}
+
 			var reportThesisDetails = new ReportThesisDetails
 			{
 				ReportThesis = reportThesis,
@@ -58,23 +71,31 @@ namespace ScientificReport.Controllers
 		// GET: ReportThesis/Create
 		public IActionResult Create()
 		{
-			return View();
+			return View(new ReportThesisModel
+			{
+				Conferences = _conferenceService.GetAll()
+			});
 		}
 
 		// POST: ReportThesis/Create
-		// To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-		// more details see http://go.microsoft.com/fwlink/?LinkId=317598.
 		[HttpPost]
 		[ValidateAntiForgeryToken]
-		public IActionResult Create([Bind("Id,Thesis")] ReportThesis reportThesis)
+		public IActionResult Create(ReportThesisModel model)
 		{
-			if (!ModelState.IsValid) return View(reportThesis);
+			if (!ModelState.IsValid)
+			{
+				model.Conferences = _conferenceService.GetAll();
+				return View(model);
+			}
 
-			_reportThesisService.CreateItem(reportThesis);
+			model.Conference = _conferenceService.GetById(model.ConferenceId);
+			_reportThesisService.CreateItem(model);
+			_reportThesisService.AddAuthor(_reportThesisService.Get(r => r.Thesis == model.Thesis).Id,
+				_userProfileService.Get(User).Id);
 			return RedirectToAction(nameof(Index));
 		}
 
-		// GET: ReportThesis/Edit/5
+		// GET: ReportThesis/Edit/{id}
 		public IActionResult Edit(Guid? id)
 		{
 			if (id == null)
@@ -88,53 +109,48 @@ namespace ScientificReport.Controllers
 				return NotFound();
 			}
 
-			var reportThesisEdt = new ReportThesisEdit
+			if (!UserHasPermission(reportThesis))
 			{
-				ReportThesis = reportThesis,
+				return Forbid();
+			}
+
+			return View(new ReportThesisEdit(reportThesis)
+			{
 				Authors = _reportThesisService.GetAuthors(reportThesis.Id),
 				Users = _userProfileService.GetAll()
-			};
-			return View(reportThesisEdt);
+			});
 		}
 
-		// POST: ReportThesis/Edit/5
-		// To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-		// more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+		// POST: ReportThesis/Edit/{id}
 		[HttpPost]
 		[ValidateAntiForgeryToken]
-		public IActionResult Edit(Guid? id, ReportThesisEdit reportThesisEdit)
+		public IActionResult Edit(Guid? id, ReportThesisEdit model)
 		{
-			var reportThesis = reportThesisEdit.ReportThesis;
+			var reportThesis = _reportThesisService.GetById(model.Id);
 			if (id != reportThesis.Id)
 			{
 				return NotFound();
 			}
+			
+			if (!UserHasPermission(reportThesis))
+			{
+				return Forbid();
+			}
 
 			if (!ModelState.IsValid)
 			{
-				reportThesisEdit.Authors = _reportThesisService.GetAuthors(reportThesis.Id);
-				reportThesisEdit.Users = _userProfileService.GetAll();
-				return View(reportThesisEdit);
+				model.Authors = _reportThesisService.GetAuthors(reportThesis.Id);
+				model.Users = _userProfileService.GetAll();
+				return View(model);
 			}
 
-			try
-			{
-				_reportThesisService.UpdateItem(reportThesis);
-			}
-			catch (DbUpdateConcurrencyException)
-			{
-				if (!_reportThesisService.Exists(reportThesis.Id))
-				{
-					return NotFound();
-				}
-
-				throw;
-			}
+			model.Conference = _conferenceService.GetById(model.ConferenceId);
+			_reportThesisService.UpdateItem(model);
 
 			return RedirectToAction(nameof(Index));
 		}
 
-		// GET: ReportThesis/Delete/5
+		// GET: ReportThesis/Delete/{id}
 		public IActionResult Delete(Guid? id)
 		{
 			if (id == null)
@@ -147,40 +163,78 @@ namespace ScientificReport.Controllers
 			{
 				return NotFound();
 			}
+			
+			if (!UserHasPermission(reportThesis))
+			{
+				return Forbid();
+			}
 
 			return View(reportThesis);
 		}
 
-		// POST: ReportThesis/Delete/5
+		// POST: ReportThesis/Delete/{id}
 		[HttpPost, ActionName("Delete")]
 		[ValidateAntiForgeryToken]
 		public IActionResult DeleteConfirmed(Guid id)
 		{
+			if (!_reportThesisService.Exists(id))
+			{
+				return NotFound();
+			}
+			
+			if (!UserHasPermission(_reportThesisService.GetById(id)))
+			{
+				return Forbid();
+			}
+			
 			_reportThesisService.DeleteById(id);
 			return RedirectToAction(nameof(Index));
 		}
 
-		private bool ReportThesisExists(Guid id)
-		{
-			return _reportThesisService.GetAll().Any(e => e.Id == id);
-		}
-
-		// POST: ScientificWork/AddAuthor/5
+		// POST: ScientificWork/AddAuthor/{id}
 		[HttpPost]
-//		[ValidateAntiForgeryToken]
 		public IActionResult AddAuthor(Guid id, [FromBody] ReportThesisAuthorRequest request)
 		{
+			if (!_reportThesisService.Exists(id))
+			{
+				return NotFound();
+			}
+			
+			if (!UserHasPermission(_reportThesisService.GetById(id)))
+			{
+				return Forbid();
+			}
+			
 			_reportThesisService.AddAuthor(id, request.AuthorId);
 			return Json(ApiResponse.Ok);
 		}
 
-		// POST: ScientificWork/DeleteAuthor/5
+		// POST: ScientificWork/DeleteAuthor/{id}
 		[HttpPost]
-//		[ValidateAntiForgeryToken]
 		public IActionResult DeleteAuthor(Guid id, [FromBody] ReportThesisAuthorRequest request)
 		{
+			if (!_reportThesisService.Exists(id))
+			{
+				return NotFound();
+			}
+			
+			if (!UserHasPermission(_reportThesisService.GetById(id)))
+			{
+				return Forbid();
+			}
+			
 			_reportThesisService.RemoveAuthor(id, request.AuthorId);
 			return Json(ApiResponse.Ok);
+		}
+		
+		private bool UserHasPermission(ReportThesis reportThesis)
+		{
+			var user = _userProfileService.Get(User);
+			var department = _departmentService.Get(d => d.Staff.Contains(user));
+			return PageHelpers.IsAdmin(User) ||
+			       PageHelpers.IsHeadOfDepartment(User) &&
+			       _reportThesisService.GetAuthors(reportThesis.Id).Any(p => department.Staff.Contains(p)) ||
+			       _reportThesisService.GetAuthors(reportThesis.Id).Contains(user);
 		}
 	}
 }
